DynamoDB/main/requirements.txt
flask
boto3
werkzeug
functools
pyjwt